+++ src/components/index.js
@@ -1,10 +1,11 @@
 import Vue from 'vue';
 
+import StBadge from './badge/index.vue';
 import StButtonGroup from './button-group/index.vue';
 import StButton from './button/index.vue';
 import StCheckbox from './checkbox/index.vue';
 import StCol from './col/index.vue';
-import Icon from './icon/index.vue';
+import StIcon from './icon/index.vue';
 import StInput from './input/index.vue';
 import StNotification from './notification/index.vue';
 import StNotificationsGroup from './notifications-group/index.vue';
@@ -17,11 +18,12 @@ import StSwitch from './switch/index.vue';
 import StTextarea from './textarea/index.vue';
 
 const Components = {
+  StBadge,
   StButton,
   StButtonGroup,
   StCheckbox,
   StCol,
-  Icon,
+  StIcon,
   StInput,
   StNotification,
   StNotificationsGroup,