+++ src/components/notifications-group/script.ts
@@ -4,6 +4,8 @@ import {
   Vue,
 } from 'vue-property-decorator';
 
+import { NotificationsGroupPosition } from './types';
+
 export const NOTIFICATIONS_GROUP_ID = 'st-notifications-group';
 
 
@@ -11,8 +13,8 @@ export const NOTIFICATIONS_GROUP_ID = 'st-notifications-group';
   name: 'StNotificationsGroup',
 })
 export default class StNotificationsGroup extends Vue {
-  @Prop({ type: String, default: 'bottom-right' })
-  position!: string;
+  @Prop({ type: String, default: NotificationsGroupPosition.bottomRight })
+  position!: NotificationsGroupPosition;
 
   get id(): string {
     return NOTIFICATIONS_GROUP_ID;