+++ src/components/notifications-group/types.ts
@@ -1,0 +1,8 @@
+export enum NotificationsGroupPosition {
+  topLeft = 'top-left',
+  topCenter = 'top-center',
+  topRight = 'top-right',
+  bottomLeft = 'bottom-left',
+  bottomCenter = 'bottom-center',
+  bottomRight = 'bottom-right',
+}