+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Changelog
 
-## 0.2.0
+## 0.2.1
 
 ### Features
 * Added `Radio` component to `Offers` theme
+* Added new component: `Switch`
 * Added new component: `Notifications`
 * Added new component: `Scrollbar`
 * Added new component: `Pagination`
+* Added new component: `Popover`
+* Added new component: `Badge`
 
 ### Maintenance
 * Fixed `Offers` theme for `Button` component